+++ backend/src/agents/datastore_agent.py
@@ -3,6 +3,7 @@ from src.utils.graph_db_utils import execute_query
 from src.prompts import PromptEngine
 from datetime import datetime
 from src.utils import to_json
+from src.utils.user_logger import user_log_info
 from .types import Parameter
 from .agent import Agent, agent
 from .tool import tool
@@ -62,11 +63,11 @@ def generate_query(question_intent, operation, question_params, aggregation, sor
     )
     llm_query = llm.chat(model, generate_cypher_query_prompt, details_to_create_cypher_query)
     json_query = to_json(llm_query)
-    logger.info(f"USER - Cypher generated by the LLM: {llm_query}")
+    user_log_info(f"Cypher generated by the LLM: {llm_query}", __name__)
     if json_query["query"] == "None":
         return "No database query"
     db_response = execute_query(json_query["query"])
-    logger.info(f"USER - Database response: {db_response}")
+    user_log_info(f"Database response: {db_response}", __name__)
     return str(db_response)
